integrate remaining pseudocode definitions docs
